+++ src/Core/Message/AbstractRequest.php
@@ -118,9 +118,14 @@ abstract class AbstractRequest extends \Omnipay\Common\Message\AbstractRequest
     {
         return $this->getParameter('rurl');
     }
-    
+
     public function setRurl($value)
     {
         return $this->setParameter('rurl', $value);
+    }
+
+    public function getHttpClient()
+    {
+        return $this->httpClient;
     }
 }